resvg::filter

Struct Image

struct Image {
    image: Rc<Pixmap>,
    region: IntRect,
    color_space: ColorInterpolation,
}

Fields
image: Rc<Pixmap>

Filter primitive result.

All images have the same size which is equal to the current filter region.

region: IntRect

Imageโs region that has actual data.

Region is in global coordinates and not in image one.

Imageโs content outside this region will be transparent/cleared.

Currently used only for feTile.

color_space: ColorInterpolation

The current color space.

Implementations

impl Image

fn from_image(image: Pixmap, color_space: ColorInterpolation) -> Self

fn into_color_space( self, color_space: ColorInterpolation, ) -> Result<Self, Error>

fn take(self) -> Result<Pixmap, Error>

fn width(&self) -> u32

fn height(&self) -> u32

fn as_ref(&self) -> &Pixmap